requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
asyncssh==2.15.0
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.137
botocore==1.34.137
bs4==0.0.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
Django==5.0.7
django-phonenumber-field==8.0.0
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.1
entrypoints==0.4
funcy==2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
gto==1.7.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
joblib==1.4.2
Mako==1.3.5
Markdown==3.6
marshmallow==3.21.3
matplotlib==3.9.0
ml-dtypes==0.3.2
mlflow==2.14.3
namex==0.0.8
numpy==1.26.4
oauth2client==4.1.3
pandas==2.2.2
prometheus_client==0.20.0
querystring-parser==1.2.4
referencing==0.35.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.14.0
setuptools==69.5.1
SQLAlchemy==2.0.31
urllib3==2.2.2
zipp==3.19.2
